Moved all the compilation stuff out of Evaluator. Things compile... but not tested
