+++ project/Constants.scala
@@ -1,5 +1,5 @@
 package ammonite
 
 object Constants{
-  val version = "0.6.0"
+  val version = "0.6.1-SNAPSHOT"
 }
