+++ repl/src/main/scala/ammonite/repl/interp/Evaluator.scala
@@ -31,39 +31,30 @@ trait Evaluator{
    * passing in the callback ensures the printing is still done lazily, but within
    * the exception-handling block of the `Evaluator`
    */
-  def processLine(code: String,
-                  printCode: String,
+  def processLine(classFiles: Util.ClassFiles,
+                  newImports: Seq[ImportData],
                   printer: Printer,
                   fileName: String,
-                  extraImports: Seq[ImportData] = Seq()
-                 ): Res[Evaluated]
+                  extraImports: Seq[ImportData] = Seq()): Res[Evaluated]
 
-  def processScriptBlock(code: String,
-                         scriptImports: Seq[ImportData],
-                         printer: Printer,
+  def processScriptBlock(cls: Class[_],
+                         newImports: Seq[ImportData],
                          wrapperName: String,
-                         fileName: String,
                          pkgName: String): Res[Evaluated]
-
-  def previousImportBlock: String
 
   def sess: Session
 
-  /*
-   * How many wrappers has this instance compiled
-   */ 
-  def compilationCount: Int
 }
 
 object Evaluator{
 
+  def interrupted(e: Throwable) = {
+    Thread.interrupted()
+    Res.Failure(Some(e), "\nInterrupted!")
+  }
 
   def apply(currentClassloader: ClassLoader,
-            compile: => (Array[Byte], Printer, Int, String) => Compiler.Output,
-            startingLine: Int,
-            cacheLoad: (String, String) => Option[CompileCache],
-            cacheSave: (String, String, CompileCache) => Unit,
-            addToCompilerClasspath:  => ClassFiles => Unit): Evaluator = new Evaluator{ eval =>
+            startingLine: Int): Evaluator = new Evaluator{ eval =>
 
     /**
      * Imports which are required by earlier commands to the REPL. Imports
@@ -140,22 +131,8 @@ object Evaluator{
       }
     }
 
-    private var _compilationCount = 0
-    def compilationCount = _compilationCount
 
-    def compileClass(code: (String, Int),
-                     printer: Printer,
-                     fileName: String): Res[(ClassFiles, Seq[ImportData])] = for {
-      compiled <- Res.Success{
-        val c = compile(code._1.getBytes, printer, code._2, fileName)
-        c
-      }
-      _ = _compilationCount += 1
-      (classfiles, imports) <- Res[(ClassFiles, Seq[ImportData])](
-        compiled,
-        "Compilation Failed"
-      )
-    } yield (classfiles, imports)
+
 
     def loadClass(fullName: String, classFiles: ClassFiles): Res[Class[_]] = {
       Res[Class[_]](Try {
@@ -172,47 +149,7 @@ object Evaluator{
     def evalMain(cls: Class[_]) =
       cls.getDeclaredMethod("$main").invoke(null)
 
-    def previousImportBlock = importBlock(imports)
 
-    def importBlock(importData: Seq[ImportData]) = {
-      Timer("importBlock 0")
-      // Group the remaining imports into sliding groups according to their
-      // prefix, while still maintaining their ordering
-      val grouped = mutable.Buffer[mutable.Buffer[ImportData]]()
-      for(data <- importData){
-        if (grouped.isEmpty) grouped.append(mutable.Buffer(data))
-        else {
-          val last = grouped.last.last
-
-          // Start a new import if we're importing from somewhere else, or
-          // we're importing the same thing from the same place but aliasing
-          // it to a different name, since you can't import the same thing
-          // twice in a single import statement
-          val startNewImport =
-            last.prefix != data.prefix || grouped.last.exists(_.fromName == data.fromName)
-
-          if (startNewImport) grouped.append(mutable.Buffer(data))
-          else grouped.last.append(data)
-        }
-      }
-      // Stringify everything
-      val out = for(group <- grouped) yield {
-        val printedGroup = for(item <- group) yield{
-          if (item.fromName == item.toName) Parsers.backtickWrap(item.fromName)
-          else s"${Parsers.backtickWrap(item.fromName)} => ${Parsers.backtickWrap(item.toName)}"
-        }
-        "import " + group.head.prefix + ".{\n  " + printedGroup.mkString(",\n  ") + "\n}\n"
-      }
-      val res = out.mkString
-
-      Timer("importBlock 1")
-      res
-    }
-
-    def interrupted(e: Throwable) = {
-      Thread.interrupted()
-      Res.Failure(Some(e), "\nInterrupted!")
-    }
 
     type InvEx = InvocationTargetException
     type InitEx = ExceptionInInitializerError
@@ -232,109 +169,38 @@ object Evaluator{
       case Ex(userEx@_*)                      =>   Res.Exception(userEx(0), "")
 
     }
-    def processLine(code: String,
-                    printCode: String,
+
+    def processLine(classFiles: Util.ClassFiles,
+                    newImports: Seq[ImportData],
                     printer: Printer,
                     fileName: String,
-                    extraImports: Seq[ImportData] = Seq()
-                   ) = for {
-      wrapperName <- Res.Success("cmd" + getCurrentLine)
-      _ <- Catching{ case e: ThreadDeath => interrupted(e) }
-      (classFiles, newImports) <- compileClass(
-        wrapCode(
-          "ammonite.session",
-          wrapperName,
-          code,
-          printCode,
-          Frame.mergeImports(imports, extraImports)
-        ),
-        printer,
-        fileName
-      )
-      _ = Timer("eval.processLine compileClass end")
-      cls <- loadClass("ammonite.session." + wrapperName, classFiles)
-      _ = Timer("eval.processLine loadClass end")
-      _ = currentLine += 1
-      _ <- Catching{userCodeExceptionHandler}
-    } yield {
-      // Exhaust the printer iterator now, before exiting the `Catching`
-      // block, so any exceptions thrown get properly caught and handled
+                    extraImports: Seq[ImportData] = Seq()) = {
+      val wrapperName = "cmd" + getCurrentLine
+      Timer("eval.processLine compileClass end")
+      for {
+        cls <- loadClass("ammonite.session." + wrapperName, classFiles)
+        _ = Timer("eval.processLine loadClass end")
+        _ = currentLine += 1
+        _ <- Catching{userCodeExceptionHandler}
+      } yield {
+        // Exhaust the printer iterator now, before exiting the `Catching`
+        // block, so any exceptions thrown get properly caught and handled
 
-      val iter = evalMain(cls).asInstanceOf[Iterator[String]]
-      Timer("eval.processLine evaluatorRunPrinter 1")
-      evaluatorRunPrinter(iter.foreach(printer.out))
-      Timer("eval.processLine evaluatorRunPrinter end")
-      evaluationResult("ammonite.session." + wrapperName, newImports)
-    }
-
-    def wrapCode(pkgName: String,
-                 indexedWrapperName: String,
-                 code: String,
-                 printCode: String,
-                 imports: Seq[ImportData]) = {
-      val topWrapper = s"""
-package $pkgName
-${importBlock(imports)}
-
-object $indexedWrapperName{\n"""
-
-     val bottomWrapper = s"""\ndef $$main() = { $printCode }
-  override def toString = "$indexedWrapperName"
-}
-"""
-     val importsLen = topWrapper.length
-     (topWrapper + code + bottomWrapper, importsLen)
-    }
-
-    def cachedCompileBlock(code: String,
-                           imports: Seq[ImportData],
-                           printer: Printer,
-                           wrapperName: String,
-                           fileName: String,
-                           pkgName: String,
-                           printCode: String = ""): Res[(Class[_], Seq[ImportData])] = {
-
-      Timer("cachedCompileBlock 1")
-
-      val wrappedCode = wrapCode(
-        pkgName, wrapperName, code, printCode, imports
-      )
-      val fullyQualifiedName = pkgName + "." + wrapperName
-      val tag = cacheTag(code, imports, sess.frames.head.classloader.classpathHash)
-      Timer("cachedCompileBlock 2")
-      val compiled = cacheLoad(fullyQualifiedName, tag) match {
-        case Some((classFiles, newImports)) =>
-          addToCompilerClasspath(classFiles)
-          Res.Success((classFiles, newImports))
-        case _ =>
-          val noneCalc = for {
-            (classFiles, newImports) <- compileClass(
-              wrappedCode, printer, fileName
-            )
-            _ = cacheSave(fullyQualifiedName, tag, (classFiles, newImports))
-          } yield (classFiles, newImports)
-
-          noneCalc
+        val iter = evalMain(cls).asInstanceOf[Iterator[String]]
+        Timer("eval.processLine evaluatorRunPrinter 1")
+        evaluatorRunPrinter(iter.foreach(printer.out))
+        Timer("eval.processLine evaluatorRunPrinter end")
+        evaluationResult("ammonite.session." + wrapperName, newImports)
       }
-      Timer("cachedCompileBlock 3")
-      val x = for {
-        (classFiles, newImports) <- compiled
-        _ = Timer("cachedCompileBlock 4")
-        cls <- loadClass(fullyQualifiedName, classFiles)
-      } yield (cls, newImports)
-
-      x
     }
 
-    def processScriptBlock(code: String,
-                           scriptImports: Seq[ImportData],
-                           printer: Printer,
+
+
+
+    def processScriptBlock(cls: Class[_],
+                           newImports: Seq[ImportData],
                            wrapperName: String,
-                           fileName: String,
                            pkgName: String) = for {
-      (cls, newImports) <- cachedCompileBlock(
-        code, scriptImports,  printer, wrapperName, fileName, pkgName
-      )
       _ <- Catching{userCodeExceptionHandler}
     } yield {
       Timer("cachedCompileBlock")
@@ -380,20 +246,6 @@ object $indexedWrapperName{\n"""
    */
   def evaluatorRunPrinter(f: => Unit) = f
 
-  /**
-   * This gives our cache tags for compile caching. The cache tags are a hash
-   * of classpath, previous commands (in-same-script), and the block-code.
-   * Previous commands are hashed in the wrapper names, which are contained 
-   * in imports, so we don't need to pass them explicitly.
-   */
-  def cacheTag(code: String, imports: Seq[ImportData], classpathHash: Array[Byte]): String = {
-    val bytes = Util.md5Hash(Iterator(
-      Util.md5Hash(Iterator(code.getBytes)),
-      Util.md5Hash(imports.iterator.map(_.toString.getBytes)),
-      classpathHash
-    ))
-    "cache" + bytes.map("%02x".format(_)).mkString //add prefix to make sure it begins with a letter
-  }
 
 
 }
